Documentation and New UT Cases are added.
