+++ spring-integration-hazelcast/src/test/java/org/springframework/integration/hazelcast/inbound/HazelcastReplicatedMapEventDrivenInboundChannelAdapterTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.onlinetechvision.integration.hazelcast.inbound;
+package org.springframework.integration.hazelcast.inbound;
 
 import javax.annotation.Resource;
 
@@ -31,53 +31,53 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import com.hazelcast.core.AbstractIMapEvent;
 import com.hazelcast.core.EntryEvent;
 import com.hazelcast.core.EntryEventType;
-import com.hazelcast.core.IMap;
+import com.hazelcast.core.ReplicatedMap;
 
 /**
- * HazelcastEventDrivenInboundChannelAdapterTest
+ * Hazelcast Replicated Map Event Driven Inbound Channe lAdapter Test
  * 
  * @author Eren Avsarogullari
  * @since 1.0.0
  *
  */
 @RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(locations={"/HazelcastDistributedMapEventDrivenInboundChannelAdapterTest-context.xml"})
+@ContextConfiguration(locations={"/HazelcastReplicatedMapEventDrivenInboundChannelAdapterTest-context.xml"})
 @DirtiesContext(classMode = ClassMode.AFTER_CLASS)
-public class HazelcastDistributedMapEventDrivenInboundChannelAdapterTest {
+public class HazelcastReplicatedMapEventDrivenInboundChannelAdapterTest {
 
 	@Autowired
-	private PollableChannel edMapChannel1;
+	private PollableChannel edReplicatedMapChannel1;
 	
 	@Autowired
-	private PollableChannel edMapChannel2;
+	private PollableChannel edReplicatedMapChannel2;
 	
 	@Autowired
-	private PollableChannel edMapChannel3;
+	private PollableChannel edReplicatedMapChannel3;
 	
 	@Autowired
-	private PollableChannel edMapChannel4;
+	private PollableChannel edReplicatedMapChannel4;
 		
 	@Resource
-	private IMap<Integer, User> edDistributedMap1;
+	private ReplicatedMap<Integer, User> edReplicatedMap1;
 	
 	@Resource
-	private IMap<Integer, User> edDistributedMap2;
+	private ReplicatedMap<Integer, User> edReplicatedMap2;
 	
 	@Resource
-	private IMap<Integer, User> edDistributedMap3;
+	private ReplicatedMap<Integer, User> edReplicatedMap3;
 	
 	@Resource
-	private IMap<Integer, User> edDistributedMap4;
+	private ReplicatedMap<Integer, User> edReplicatedMap4;
 	
 	@Test
 	public void testEventDrivenForOnlyADDEDEntryEvent() {
-		edDistributedMap1.put(1, new User(1, "TestName1", "TestSurname1"));
-		Message<?> msg = edMapChannel1.receive(2_000);
+		edReplicatedMap1.put(1, new User(1, "TestName1", "TestSurname1"));
+		Message<?> msg = edReplicatedMapChannel1.receive(2_000);
 		Assert.assertNotNull(msg);
 		Assert.assertNotNull(msg.getPayload());
 		Assert.assertTrue(msg.getPayload() instanceof EntryEvent);
 		Assert.assertEquals(EntryEventType.ADDED, ((EntryEvent<?, ?>)msg.getPayload()).getEventType());
-		Assert.assertEquals("edDistributedMap1", ((EntryEvent<?, ?>)msg.getPayload()).getName());
+		Assert.assertEquals("edReplicatedMap1", ((EntryEvent<?, ?>)msg.getPayload()).getName());
 		Assert.assertEquals(1, ((EntryEvent<?, ?>)msg.getPayload()).getKey());
 		Assert.assertEquals(1, ((User)((EntryEvent<?, ?>)msg.getPayload()).getValue()).getId());
 		Assert.assertEquals("TestName1", ((User)((EntryEvent<?, ?>)msg.getPayload()).getValue()).getName());
@@ -86,14 +86,14 @@ public class HazelcastDistributedMapEventDrivenInboundChannelAdapterTest {
 	
 	@Test
 	public void testEventDrivenForOnlyUPDATEDEntryEvent() {
-		edDistributedMap2.put(2, new User(1, "TestName1", "TestSurname1"));
-		edDistributedMap2.put(2, new User(2, "TestName2", "TestSurname2"));
-		Message<?> msg = edMapChannel2.receive(2_000);
+		edReplicatedMap2.put(2, new User(1, "TestName1", "TestSurname1"));
+		edReplicatedMap2.put(2, new User(2, "TestName2", "TestSurname2"));
+		Message<?> msg = edReplicatedMapChannel2.receive(2_000);
 		Assert.assertNotNull(msg);
 		Assert.assertNotNull(msg.getPayload());
 		Assert.assertTrue(msg.getPayload() instanceof EntryEvent);
 		Assert.assertEquals(EntryEventType.UPDATED, ((EntryEvent<?, ?>)msg.getPayload()).getEventType());
-		Assert.assertEquals("edDistributedMap2", ((EntryEvent<?, ?>)msg.getPayload()).getName());
+		Assert.assertEquals("edReplicatedMap2", ((EntryEvent<?, ?>)msg.getPayload()).getName());
 		Assert.assertEquals(2, ((EntryEvent<?, ?>)msg.getPayload()).getKey());
 		Assert.assertEquals(1, ((User)((EntryEvent<?, ?>)msg.getPayload()).getOldValue()).getId());
 		Assert.assertEquals("TestName1", ((User)((EntryEvent<?, ?>)msg.getPayload()).getOldValue()).getName());
@@ -105,15 +105,15 @@ public class HazelcastDistributedMapEventDrivenInboundChannelAdapterTest {
 	
 	@Test
 	public void testEventDrivenForOnlyREMOVEDEntryEvent() {
-		edDistributedMap3.put(1, new User(1, "TestName1", "TestSurname1"));
-		edDistributedMap3.put(2, new User(2, "TestName2", "TestSurname2"));
-		edDistributedMap3.remove(2);
-		Message<?> msg = edMapChannel3.receive(2_000);
+		edReplicatedMap3.put(1, new User(1, "TestName1", "TestSurname1"));
+		edReplicatedMap3.put(2, new User(2, "TestName2", "TestSurname2"));
+		edReplicatedMap3.remove(2);
+		Message<?> msg = edReplicatedMapChannel3.receive(2_000);
 		Assert.assertNotNull(msg);
 		Assert.assertNotNull(msg.getPayload());
 		Assert.assertTrue(msg.getPayload() instanceof EntryEvent);
 		Assert.assertEquals(EntryEventType.REMOVED, ((EntryEvent<?, ?>)msg.getPayload()).getEventType());
-		Assert.assertEquals("edDistributedMap3", ((EntryEvent<?, ?>)msg.getPayload()).getName());
+		Assert.assertEquals("edReplicatedMap3", ((EntryEvent<?, ?>)msg.getPayload()).getName());
 		Assert.assertEquals(2, ((EntryEvent<?, ?>)msg.getPayload()).getKey());
 		Assert.assertEquals(2, ((User)((EntryEvent<?, ?>)msg.getPayload()).getOldValue()).getId());
 		Assert.assertEquals("TestName2", ((User)((EntryEvent<?, ?>)msg.getPayload()).getOldValue()).getName());
@@ -122,25 +122,22 @@ public class HazelcastDistributedMapEventDrivenInboundChannelAdapterTest {
 	
 	@Test
 	public void testEventDrivenForALLEntryEvent() {
-		edDistributedMap4.put(1, new User(1, "TestName1", "TestSurname1"));
-		Message<?> msg = edMapChannel4.receive(2_000);
-		verify(msg, "edDistributedMap4", EntryEventType.ADDED);
+		edReplicatedMap4.put(1, new User(1, "TestName1", "TestSurname1"));
+		Message<?> msg = edReplicatedMapChannel4.receive(2_000);
+		verify(msg, "edReplicatedMap4", EntryEventType.ADDED);
 		
-		edDistributedMap4.put(1, new User(1, "TestName1", "TestSurnameUpdated"));
-		msg = edMapChannel4.receive(2_000);
-		verify(msg, "edDistributedMap4", EntryEventType.UPDATED);
+		edReplicatedMap4.put(1, new User(1, "TestName1", "TestSurnameUpdated"));
+		msg = edReplicatedMapChannel4.receive(2_000);
+		verify(msg, "edReplicatedMap4", EntryEventType.UPDATED);
 		
-		edDistributedMap4.remove(1);
-		msg = edMapChannel4.receive(2_000);
-		verify(msg, "edDistributedMap4", EntryEventType.REMOVED);
+		edReplicatedMap4.remove(1);
+		msg = edReplicatedMapChannel4.receive(2_000);
+		verify(msg, "edReplicatedMap4", EntryEventType.REMOVED);
 		
-		edDistributedMap4.put(2, new User(2, "TestName2", "TestSurname2"));
-		msg = edMapChannel4.receive(2_000);
-		verify(msg, "edDistributedMap4", EntryEventType.ADDED);
+		edReplicatedMap4.put(2, new User(2, "TestName2", "TestSurname2"));
+		msg = edReplicatedMapChannel4.receive(2_000);
+		verify(msg, "edReplicatedMap4", EntryEventType.ADDED);
 		
-		edDistributedMap4.clear();
-		msg = edMapChannel4.receive(2_000);
-		verify(msg, "edDistributedMap4", EntryEventType.CLEAR_ALL);
 	}
 		
 	private void verify(Message<?> msg, String cacheName, EntryEventType type) {
@@ -150,7 +147,5 @@ public class HazelcastDistributedMapEventDrivenInboundChannelAdapterTest {
 		Assert.assertEquals(cacheName, ((AbstractIMapEvent)msg.getPayload()).getName());
 		Assert.assertEquals(type, ((AbstractIMapEvent)msg.getPayload()).getEventType());
 	}
-	
-	
-	
+
 }
